requirements.txt
fastapi==0.125.0
uvicorn==0.38.0
psycopg==3.3.2
requests==2.32.3
PyYAML==6.0.3